Removed remaining duplicate valueName parameters
Removed duplicate 'valueName' attributes from all remaining Policy nodes where child Policy Element nodes also have a 'valueName' attribute defined and both values are equal (example: both have an attribute of valueName="foo").

The valueName attribute on the Policy node specifies the registry value to check to determine whether the policy is "Disabled" (0), "Enabled" (non-zero), or "Not Configured" (missing).  A non-numeric value here will cause the policy to be evaluated as "Not Configured".

The valueName attribute on the Policy Element nodes specifies the registry value where the user's selections should be stored.

When the valueName for the Policy and Policy Element nodes are set to the same value, it results in unexpected behavior in certain cases.  For example, an enum with a decimal value of 0 would cause the policy to be evaluated as "Disabled" or a text value (non-numeric) will cause the policy to be evaluated as "Not Configured" even though the expected value may be present in the registry.